extract method and change error name and message
