+++ components-api/src/main/scala/pl/touk/nussknacker/engine/api/context/ProcessCompilationError.scala
@@ -257,13 +257,13 @@ object ProcessCompilationError {
       OverwrittenVariable(variableName, nodeId.id, paramName)
   }
 
-  final case class InvalidVariableOutputName(name: String, nodeId: String, paramName: Option[String])
+  final case class InvalidVariableName(name: String, nodeId: String, paramName: Option[String])
       extends PartSubGraphCompilationError
       with InASingleNode
 
-  object InvalidVariableOutputName {
+  object InvalidVariableName {
     def apply(variableName: String, paramName: Option[String])(implicit nodeId: NodeId): PartSubGraphCompilationError =
-      InvalidVariableOutputName(variableName, nodeId.id, paramName)
+      InvalidVariableName(variableName, nodeId.id, paramName)
   }
 
   final case class FragmentOutputNotDefined(id: String, nodeIds: Set[String]) extends ProcessCompilationError
