+++ components-api/src/main/scala/pl/touk/nussknacker/engine/api/validation/Validations.scala
@@ -4,7 +4,7 @@ import cats.data.Validated.{Invalid, Valid}
 import cats.data.ValidatedNel
 import pl.touk.nussknacker.engine.api.NodeId
 import pl.touk.nussknacker.engine.api.context.PartSubGraphCompilationError
-import pl.touk.nussknacker.engine.api.context.ProcessCompilationError.InvalidVariableOutputName
+import pl.touk.nussknacker.engine.api.context.ProcessCompilationError.InvalidVariableName
 
 import javax.lang.model.SourceVersion
 
@@ -15,7 +15,7 @@ object Validations {
   ): ValidatedNel[PartSubGraphCompilationError, String] = {
     // TODO: add correct and more precise error messages
     if (isVariableNameValid(name)) Valid(name)
-    else Invalid(InvalidVariableOutputName(name, paramName)).toValidatedNel
+    else Invalid(InvalidVariableName(name, paramName)).toValidatedNel
   }
 
   def isVariableNameValid(name: String): Boolean = SourceVersion.isIdentifier(name)
